+++ packages/ebsr/configure/src/main.jsx
@@ -29,7 +29,7 @@ export class Main extends React.Component {
   removeExtraChoices = (choices) => {
     let correctFound = false;
 
-    return (choices || []).map(choice => {
+     return (choices || []).map(choice => {
       if (correctFound) {
         choice.correct = false;
         return choice;
@@ -68,7 +68,7 @@ export class Main extends React.Component {
       configuration,
       onConfigurationChanged
     } = this.props;
-    const { partLabelType, partA: modelPartA, partB: modelPartB } = model;
+    const { partLabelType } = model;
     const { partA, partB, partialScoring, scoringType, ...generalConfiguration } = configuration;
     const {
       feedback: feedbackA = {},
@@ -78,8 +78,7 @@ export class Main extends React.Component {
       prompt: promptA = {},
       teacherInstructions: teacherInstructionsA = {},
       studentInstructions: studentInstructionsA = {},
-      choicesLayout: choicesLayoutA = {},
-      gridColumns: gridColumnsA = {},
+      verticalMode: verticalModeA = {},
       rationale: rationaleA = {}
     } = partA || {};
     const {
@@ -90,20 +89,13 @@ export class Main extends React.Component {
       prompt: promptB = {},
       teacherInstructions: teacherInstructionsB = {},
       studentInstructions: studentInstructionsB = {},
-      choicesLayout: choicesLayoutB = {},
-      gridColumns: gridColumnsB = {},
+      verticalMode: verticalModeB = {},
       rationale: rationaleB = {}
     } = partB || {};
     const type = partLabelType || 'Numbers';
     const typeIsNumber = type === 'Numbers';
     const firstPart = `Part ${typeIsNumber ? '1' : 'A'}`;
     const secondPart = `Part ${typeIsNumber ? '2' : 'B'}`;
-    const nrOfColumnsAvailable = {
-      partA: modelPartA.choices && modelPartA.choices.length ?
-        Array.from({length: modelPartA.choices.length}, (_, i) => (`${i + 1}`)) : [],
-      partB: modelPartB.choices && modelPartB.choices.length ?
-        Array.from({length: modelPartB.choices.length}, (_, i) => (`${i + 1}`)) : [],
-    };
 
     return (
       <div className={classes.design}>
@@ -135,12 +127,8 @@ export class Main extends React.Component {
                     radio(choicePrefixA.label, ['numbers', 'letters']),
                   'partA.lockChoiceOrder':
                     lockChoiceOrderA.settings && toggle(lockChoiceOrderA.label),
-                  'partA.choicesLayout':
-                    choicesLayoutA.settings &&
-                    dropdown(choicesLayoutA.label, ['vertical', 'grid', 'horizontal']),
-                  'partA.gridColumns':
-                    modelPartA.choicesLayout === 'grid' && nrOfColumnsAvailable.partA.length > 0 &&
-                    dropdown(gridColumnsA.label, nrOfColumnsAvailable.partA)
+                  'partA.verticalMode':
+                    verticalModeA.settings && toggle(verticalModeA.label)
                 },
                 [`Properties ${firstPart}`]: {
                   'partA.feedbackEnabled': feedbackA.settings &&
@@ -165,12 +153,8 @@ export class Main extends React.Component {
                     radio(choicePrefixB.label, ['numbers', 'letters']),
                   'partB.lockChoiceOrder':
                     lockChoiceOrderB.settings && toggle(lockChoiceOrderB.label),
-                  'partB.choicesLayout':
-                    choicesLayoutB.settings &&
-                    dropdown(choicesLayoutB.label, ['vertical', 'grid', 'horizontal']),
-                  'partB.gridColumns':
-                    modelPartB.choicesLayout === 'grid' && nrOfColumnsAvailable.partB.length > 0 &&
-                    dropdown(gridColumnsB.label, nrOfColumnsAvailable.partB)
+                  'partB.verticalMode':
+                    verticalModeB.settings && toggle(verticalModeB.label)
                 },
                 [`Properties ${secondPart}`]: {
                   'partB.feedbackEnabled':