+++ packages/graphing/configure/src/defaults.js
@@ -1,6 +1,6 @@
-import { tools } from '@pie-lib/graphing';
+import {tools} from '@pie-lib/graphing';
 
-const { allTools = [] } = tools;
+const {allTools = []} = tools;
 /**
  * NOTE: There's no functionality described for padding
  * so there's no implementation (they are only added in model)
@@ -9,7 +9,12 @@ const { allTools = [] } = tools;
 export default {
   model: {
     answers: {},
-    arrows: true,
+    arrows: {
+      left: true,
+      right: true,
+      up: true,
+      down: true
+    },
     backgroundMarks: [],
     domain: {
       min: -5,
@@ -35,6 +40,7 @@ export default {
     rationale: '',
     title: '',
     toolbarTools: allTools,
+    coordinatesOnHover: false,
     promptEnabled: true,
     rationaleEnabled: true,
     teacherInstructionsEnabled: true,
@@ -47,8 +53,20 @@ export default {
       enabled: false
     },
     arrows: {
-      settings: false,
-      label: 'Include arrows'
+      settings: true,
+      label: 'Include Arrows',
+      left: {
+        label: 'left'
+      },
+      right: {
+        label: 'right'
+      },
+      up: {
+        label: 'up'
+      },
+      down: {
+        label: 'down'
+      }
     },
     padding: {
       settings: false,
@@ -83,6 +101,10 @@ export default {
       settings: false,
       label: 'Graph Title',
       enabled: true
+    },
+    coordinatesOnHover: {
+      settings: true,
+      label: 'Coordinates on Hover'
     }
   }
 };