+++ packages/inline-dropdown/configure/src/defaults.js
@@ -10,7 +10,8 @@ export default {
     promptEnabled: true,
     rationaleEnabled: true,
     teacherInstructionsEnabled: true,
-    studentInstructionsEnabled: true
+    studentInstructionsEnabled: true,
+    toolbarEditorPosition: 'bottom'
   },
   configuration: {
     prompt: {
@@ -32,6 +33,10 @@ export default {
     teacherInstructions: {
       settings: true,
       label: 'Teacher Instructions'
-    }
+    },
+    toolbarEditorPosition:{
+      settings: false,
+      label: 'Toolbar Editor Position'
+    },
   }
 };