+++ packages/math-inline/controller/src/defaults.js
@@ -17,6 +17,7 @@ export default {
       type: 'none'
     }
   },
+  note: 'The answer shown above is the primary correct answer specified by the author for this item, but other answers may also be recognized as correct.',
   equationEditor: '3',
   expression: '',
   question: '',