+++ packages/charting/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.28](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting-configure@2.1.27...@pie-element/charting-configure@2.1.28) (2021-08-05)
+
+**Note:** Version bump only for package @pie-element/charting-configure
+
+
+
+
+
 ## [2.1.27](https://github.com/pie-framework/pie-elements/compare/@pie-element/charting-configure@2.1.25...@pie-element/charting-configure@2.1.27) (2021-08-05)