+++ packages/ebsr/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.7.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr@6.6.2...@pie-element/ebsr@6.7.0) (2021-08-05)
+
+
+### Features
+
+* **ebsr:** Allowed Part A to be multiple select PD-259 ([85530fc](https://github.com/pie-framework/pie-elements/commit/85530fce124ca5b09c024959c0147ac56b027a15))
+
+
+
+
+
 ## [6.6.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/ebsr@6.6.0...@pie-element/ebsr@6.6.2) (2021-08-05)