+++ packages/ruler/configure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.27](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.7.26...@pie-element/ruler-configure@2.7.27) (2021-08-05)
+
+**Note:** Version bump only for package @pie-element/ruler-configure
+
+
+
+
+
 ## [2.7.26](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.7.24...@pie-element/ruler-configure@2.7.26) (2021-08-05)